Comments pt 2 and try fixing spec test
